Cleanup Opp session during ACL disconnect.
Use Case:
1) Pair with BT and dongle and start receiving file
2) Reset BT dongle abruptly such that ACL disconnects and file transfer fails.

Failure:
File not sent error is displayed after long time.

Fix:
Process ACL_DISCONNECT message in OPP module to clean up
the transfer related data as soon as ACL is disconnected.
This removes the delay we observe in displaying File not received
message when Remote device's Battery is abbruptly taken out
or remote disconnects the ACL when authorization request is in
pending state in DUT.

CRs-fixed: 520411

Change-Id: Ie6af1ab8c354f8e21eb56c61e5518b83b621c56d