Do not delay broadcast of intermediate states.
In connection race conditions there can be delay of
1000 milliseconds for intermediate states, in such
cases we broadcast stable state without any delay
and intermediate with delay. This causes UI to stall
in intermediate state.

With this change we will make sure that there is no delay
for intermediate states so that we always broadcast stable
state as last connection state.

Bug: 22102403

Change-Id: Ia8a8291baa1136be3da38ff6d80f80fd53bca164